+++ tcod/ec/test_ec.py
@@ -83,6 +83,8 @@ def test_ComponentDict() -> None:
     assert set(entity) == entity.keys() == {Base, Foo}
     assert set(entity.values()) == {derived, foo}
     assert set(zip(entity.keys(), entity.values())) == set(entity.items())
+    assert entity == entity
+    assert entity != tcod.ec.ComponentDict(entity.values())
 
     entity = tcod.ec.ComponentDict()
     assert Base not in entity